refactor(foundation-demo): remove old daff-header component from src
