--- apps/foundation-demo/src/app/core/header/components/header-view/header-view.component.html
@@ -1,3 +1,0 @@
-<div sidebar-container #SidebarContainer="SidebarContainer">
-  <header (toggleSidebarVisibility)="SidebarContainer.toggleSidebarVisibility()"></header>
-</div>
