--- apps/foundation-demo/src/app/core/header/components/header-view/header-view.component.ts
@@ -1,7 +1,0 @@
-import { Component } from '@angular/core';
-
-@Component({
-  selector: 'header-view',
-  templateUrl: './header-view.component.html'
-})
-export class HeaderViewComponent { }
