+++ apps/foundation-demo/src/app/core/header/containers/header/header.component.html
@@ -1,0 +1,7 @@
+<header>
+    <button header-menu-button (click)="openSidebar()"><i class="fa fa-bars"></i></button>
+    <a header-logo routerLink="/">
+        <logo></logo>
+    </a>
+    <button header-cart-button><i class="fa fa-shopping-cart"></i></button>
+</header>
