+++ apps/foundation-demo/src/app/core/header/containers/header/header.component.ts
@@ -1,0 +1,18 @@
+import { Component, OnInit } from '@angular/core';
+import { Router } from '@angular/router';
+
+import { Store } from '@ngrx/store';
+
+import { OpenSidebar } from '../../../sidebar/actions/sidebar.actions';
+
+@Component({
+  selector: 'header-container',
+  templateUrl: './header.component.html'
+})
+export class HeaderContainer {
+  constructor(private store: Store<{}>) { }
+
+  openSidebar() {
+    this.store.dispatch(new OpenSidebar());
+  }
+}
