--- apps/foundation-demo/src/app/design/molecules/daff-header/daff-header.module.ts
@@ -1,13 +1,0 @@
-import { NgModule } from '@angular/core';
-
-import { DaffHeaderComponent } from './daff-header/daff-header.component';
-
-@NgModule({
-  declarations: [
-    DaffHeaderComponent
-  ],
-  exports: [
-    DaffHeaderComponent
-  ]
-})
-export class DaffHeaderModule { }
