--- apps/foundation-demo/src/app/design/molecules/daff-header/daff-header/daff-header.component.scss
@@ -1,7 +1,0 @@
-@import 'daffodil-util';
-
-.daff-header {
-  height: 70px;
-  width: 100%;
-  @include section-divider();
-}
